move dashboard placeholder content into its own component
